requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
motor
pymongo
PyPDF2
mistralai
python-multipart